Made changes to navbar, removed fake data from weight, finished add weight.
